Rename test framework files
