+++ .ghci
@@ -1,5 +1,5 @@
 :set -isrc
-:l src/Course.hs
+:l src/Test/Loader.hs
 :set prompt "λ> "
 :set -Wall
 :set -fno-warn-unused-binds
